Refactor connection state
Until now, connection state was managed in VncClient. This was fine before
introduction of SSH Tunnel, but now, state in VncClient may not fully reflect
the state transitions. Moreover, most of the time we simply check whether
we are connected or not.

Now, simple boolean state is managed by VncClient, and more granular state
is managed by VncViewModel.