Fix LifecycleOwner leak in LiveEvent
LiveEvent was keeping reference to the observers in `wrappedObservers`.
Those observers would most likely keep references to Activities/Fragments.
These references were not cleared after LifecycleOwner was destroyed.

The fix is to remove `wrappedObservers` altogether. It was only needed to
support 'observerForever()' which was only used in tests. If we ever need this,
it can be implemented by registering a lifecycle observer which clears
old event observers.